backend/requirements.txt
flask==2.0.1
tensorflow==2.8.0
pillow==9.0.0
numpy==1.21.5
gunicorn==20.1.0
flask-cors==3.0.10
Werkzeug==2.0.1 